requirements.txt
python-dateutil
torch
torchvision
torchaudio
fastapi

Pillow
pytesseract